+++ test/Braintree.Tests.Integration/CreditCardIntegrationTest.cs
@@ -1323,12 +1323,12 @@ namespace Braintree.Tests.Integration
 
           CreditCardVerification verification = card.Verification;
 
-          Assert.AreEqual("Y", verification.ThreeDSecureInfo.Enrolled);
-          Assert.AreEqual("cavv_value", verification.ThreeDSecureInfo.Cavv);
-          Assert.AreEqual("05", verification.ThreeDSecureInfo.EciFlag);
           Assert.AreEqual("authenticate_successful", verification.ThreeDSecureInfo.Status);
-          Assert.AreEqual("1.0.2", verification.ThreeDSecureInfo.ThreeDSecureVersion);
-          Assert.AreEqual("xid_value", verification.ThreeDSecureInfo.Xid);
+          Assert.IsNotNull(verification.ThreeDSecureInfo.Enrolled);
+          Assert.IsNotNull(verification.ThreeDSecureInfo.Cavv);
+          Assert.IsNotNull(verification.ThreeDSecureInfo.EciFlag);
+          Assert.IsNotNull(verification.ThreeDSecureInfo.ThreeDSecureVersion);
+          Assert.IsNotNull(verification.ThreeDSecureInfo.Xid);
           Assert.IsTrue(verification.ThreeDSecureInfo.LiabilityShifted);
           Assert.IsTrue(verification.ThreeDSecureInfo.LiabilityShiftPossible);
         }